migrating from javascript to htmx
