+++ chirps/asset/tests.py
@@ -34,8 +34,6 @@ class AssetTests(TestCase):
 
             self.assertRedirects(response, '/', 302)
 
-        self.get_embedding_models_url = reverse('get_embedding_models')
-
     def test_asset_tenant_isolation(self):
         """Verify that assets are isolated to a single tenant."""
         # Create an asset for user1
@@ -125,26 +123,6 @@ class AssetTests(TestCase):
 
         response = self.client.post(reverse('asset_create', args=['Pinecone']), form_data)
         self.assertRedirects(response, reverse('asset_dashboard'))
-
-    def test_get_embedding_models_view(self):
-        """Ensure expected models are returned"""
-        self.client.post(
-            reverse('login'),
-            {
-                'username': self.users[0]['username'],
-                'password': self.users[0]['password'],
-            },
-        )
-        service = Embedding.Service.OPEN_AI
-        response = self.client.get(self.get_embedding_models_url, {'service': service})
-
-        self.assertEqual(response.status_code, 200)
-
-        models = Embedding.get_models_for_service(service)
-        returned_models = response.json()
-
-        self.assertEqual(len(models), len(returned_models))
-        self.assertEqual({model[0] for model in models}, {model[0] for model in returned_models})
 
 
 class AssetPaginationTests(TestCase):
