+++ chirps/embedding/templates/embedding/model_choices.html
@@ -1,0 +1,3 @@
+{% for model in service_model_choices %}
+    <option value="{{model}}">{{model}}</option>
+{% endfor %}
